% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_picture.R
\name{convert_to_webm}
\alias{convert_to_webm}
\title{Converts a given video file to webm format}
\usage{
convert_to_webm(path)
}
\arguments{
\item{path}{(character) Local file path where the video file is located.}
}
\description{
Converts a given video file to webm format
}
