% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_picture.R
\name{+.teo_picture}
\alias{+.teo_picture}
\title{Arithmetic operator + for picture tokens}
\usage{
\method{+}{teo_picture}(e1, e2)
}
\arguments{
\item{e1}{(teo_picture) The first picture.}

\item{e2}{(teo_picture) The second picture.}
}
\value{
A new token that will filter images based on both pictures.
}
\description{
Arithmetic operator + for picture tokens
}
