% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_picture.R
\name{print.teo_picture}
\alias{print.teo_picture}
\title{Transforms a \code{teo} package function return value into a picture in the viewer}
\usage{
\method{print}{teo_picture}(x, ...)
}
\arguments{
\item{x}{(teo_picture class) An object returned by the other teo package functions.}
}
\description{
Transforms a \code{teo} package function return value into a picture in the viewer
}
