% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_picture.R
\name{save_content}
\alias{save_content}
\title{Saves to the image database a given image}
\usage{
save_content(path, type = "teo")
}
\arguments{
\item{path}{(character) A local path or URL where the image can be found}
}
\description{
Saves to the image database a given image
}
