% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_picture.R
\name{token}
\alias{token}
\title{Creates a token symbol that would display as a picture in the viewer}
\usage{
token(names)
}
\description{
The token names are used to filter through picture names before randomly
selecting one for display.

@param names (character) The token names.

@return A toe_picture class object that can render in the RStudio viewer.
}
